varsity_plug/requirements.txt
asgiref==3.8.1
Django==5.2
packaging==24.2
pillow==11.1.0
PyPDF2==3.0.1
pytesseract==0.3.13
sqlparse==0.5.3
tzdata==2025.2
